+++ src/Oro/Bundle/ApiBundle/Processor/Shared/AddMetaPropertyFilter.php
@@ -10,9 +10,11 @@ use Oro\Component\ChainProcessor\ContextInterface;
 use Oro\Component\ChainProcessor\ProcessorInterface;
 
 /**
- * Adds "meta" filter that can be used to specify which entity meta properties should be returned.
+ * Adds "meta" filter that can be used to specify which entity meta properties should be returned
+ * or which additional operations should be performed.
  * As this filter has influence on the entity configuration, it is handled by a separate processor.
  * @see \Oro\Bundle\ApiBundle\Processor\Shared\HandleMetaPropertyFilter
+ * @see \Oro\Bundle\ApiBundle\Processor\Shared\ValidateMetaPropertyFilterSupported
  */
 class AddMetaPropertyFilter implements ProcessorInterface
 {
@@ -45,14 +47,9 @@ class AddMetaPropertyFilter implements ProcessorInterface
             return;
         }
 
-        $config = $context->getConfig();
-        if (null === $config || !$config->isMetaPropertiesEnabled()) {
-            // the "meta" filter is disabled
-            return;
-        }
-
         $filter = new MetaPropertyFilter(DataType::STRING, self::FILTER_DESCRIPTION);
         $filter->setArrayAllowed(true);
-        $filters->add($filterName, $filter);
+        $filter->addAllowedMetaProperty('title', DataType::STRING);
+        $filters->add($filterName, $filter, false);
     }
 }