+++ CHANGELOG.md
@@ -2,7 +2,8 @@ Please refer first to [UPGRADE.md](UPGRADE.md) for the most important items that
 
 The current file describes significant changes in the code that may affect the upgrade of your customizations.
 
-## 4.1.0
+## 4.1.0 (2020-01-31)
+[Show detailed list of changes](incompatibilities-4-1.md)
 
 ### Changed
 
@@ -40,6 +41,11 @@ The current file describes significant changes in the code that may affect the u
 * Added *MultiImage* and *MultiField* field types to Entity Manager. Read more in [documentation](./src/Oro/Bundle/AttachmentBundle/README.md).
 
 ### Removed
+* `*.class` parameters for all entities were removed from the dependency injection container.
+The entity class names should be used directly, e.g. `'Oro\Bundle\EmailBundle\Entity\Email'`
+instead of `'%oro_email.email.entity.class%'` (in service definitions, datagrid config files, placeholders, etc.), and
+`\Oro\Bundle\EmailBundle\Entity\Email::class` instead of `$container->getParameter('oro_email.email.entity.class')`
+(in PHP code).
 
 #### ActivityListBundle
 * The `getActivityClass()` method was removed from `Oro\Bundle\ActivityListBundle\Model\ActivityListProviderInterface`.
@@ -182,7 +188,7 @@ The current file describes significant changes in the code that may affect the u
 ### Removed
 
 #### All Bundles
-* All `*.class` parameters were removed from the dependency injection container.
+* All `*.class` parameters for service definitions were removed from the dependency injection container.
 
 #### Math component
 * The deprecated method `Oro\Component\Math\BigDecimal::withScale()` was removed. Use `toScale()` method instead.